refactor(api): décorateurs d’authentification HTTP dans ApiClient/Auth
